Create Day 43
I needed a bit of help with the pp() subroutine just with getting the bingo card printing out how I wanted it to look and used ai to help me with figuring that out. The lists being nested is a bit confusing but with practice I should get used to it.